Replace type Int? by Double?

+++ app/src/main/java/com/example/hellocowcow/datas/response/xoxnoApi/NftValueResponse.kt
@@ -7,7 +7,7 @@ data class NftValueResponse (
 
     @SerializedName("floorValue"   ) var floorValue   : Double? = null,
     @SerializedName("avgValue"     ) var avgValue     : Double? = null,
-    @SerializedName("maxValue"     ) var maxValue     : Int?    = null,
+    @SerializedName("maxValue"     ) var maxValue     : Double?    = null,
     @SerializedName("collectionFp" ) var collectionFp : Double? = null
 
 )
